Update to the reset state to fix the animation preview problem
