+++ package.json
@@ -20,7 +20,7 @@
     "jszip": "^3.10.1",
     "nanoid": "^5.1.5",
     "route-engine-js": "0.1.8",
-    "route-graphics": "../route-graphics",
+    "route-graphics": "0.0.24",
     "rxjs": "^7.8.2",
     "snabbdom": "^3.6.2",
     "snabbdom-to-html": "^7.1.0",
