+++ src/pages/tweens/tweens.store.js
@@ -1,4 +1,5 @@
 import { toFlatGroups, toFlatItems } from "insieme";
+import { resetState } from "./constants";
 
 const form = {
   fields: [
@@ -482,40 +483,12 @@ export const selectIsGraphicsServiceInitialized = ({ state }) => {
   return state.isGraphicsServiceInitialized;
 };
 
-// Constants for preview rendering
-const CANVAS_WIDTH = 1920;
-const CANVAS_HEIGHT = 1080;
-const BG_COLOR = "#4a4a4a";
-const PREVIEW_RECT_WIDTH = 200;
-const PREVIEW_RECT_HEIGHT = 200;
 
 // Helper function to create render state with animations
 export const createAnimationRenderState = (
   properties,
   includeAnimations = true,
 ) => {
-  const elements = [
-    {
-      id: "bg",
-      type: "rect",
-      x: 0,
-      y: 0,
-      width: CANVAS_WIDTH,
-      height: CANVAS_HEIGHT,
-      fill: BG_COLOR,
-    },
-    {
-      id: "preview-element",
-      type: "rect",
-      x: CANVAS_WIDTH / 2,
-      y: CANVAS_HEIGHT / 2,
-      width: PREVIEW_RECT_WIDTH,
-      height: PREVIEW_RECT_HEIGHT,
-      fill: "white",
-      anchorX: 0.5,
-      anchorY: 0.5,
-    },
-  ];
 
   const animations = [];
   if (includeAnimations && properties && Object.keys(properties).length > 0) {
@@ -571,7 +544,7 @@ export const createAnimationRenderState = (
   }
 
   return {
-    elements,
+    ...resetState,
     animations,
   };
 };
